Run stress tests in VM
